requirements.txt
by~=0.0.7
allure-pytest
selene==2.0.0rc9
pytest==9.0.2
selenium==4.39.0